service/cloudhsmv2: Modernize and refactor resources and datasource
Reference: https://github.com/hashicorp/terraform-provider-aws/issues/18591

Changes:

* Fix filenames to match resource naming
* Refactor to use finder and waiter package patterns
* Refactor internal failure retry handling to service client level
* Additional acceptance testing and removed resource handling

Output from acceptance testing (reduced concurrency to 5 cluster limit):

```
--- PASS: TestAccAWSCloudHsmV2Cluster_basic (254.16s)
--- PASS: TestAccAWSCloudHsmV2Cluster_disappears (285.53s)
--- PASS: TestAccAWSCloudHsmV2Cluster_Tags (322.03s)

--- PASS: TestAccAWSCloudHsmV2Hsm_AvailabilityZone (831.81s)
--- PASS: TestAccAWSCloudHsmV2Hsm_basic (898.92s)
--- PASS: TestAccAWSCloudHsmV2Hsm_disappears (856.53s)
--- PASS: TestAccAWSCloudHsmV2Hsm_disappears_Cluster (903.49s)
--- PASS: TestAccAWSCloudHsmV2Hsm_IpAddress (861.12s)
```

Output from sweeper in AWS Commercial:

```
2021/04/06 15:16:30 [DEBUG] Running Sweepers for region (us-west-2):
2021/04/06 15:16:30 [DEBUG] Running Sweeper (aws_cloudhsm_v2_cluster) in region (us-west-2)
2021/04/06 15:17:02 Sweeper Tests ran successfully:
        - aws_cloudhsm_v2_cluster
2021/04/06 15:17:02 [DEBUG] Running Sweepers for region (us-east-1):
2021/04/06 15:17:02 [DEBUG] Running Sweeper (aws_cloudhsm_v2_cluster) in region (us-east-1)
2021/04/06 15:17:03 Sweeper Tests ran successfully:
        - aws_cloudhsm_v2_cluster
ok      github.com/terraform-providers/terraform-provider-aws/aws       35.345s
```
