+++ aws/resource_aws_cloudhsm_v2_cluster.go
@@ -7,10 +7,11 @@ import (
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/cloudhsmv2"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/terraform-providers/terraform-provider-aws/aws/internal/keyvaluetags"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudhsmv2/finder"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudhsmv2/waiter"
 )
 
 func resourceAwsCloudHsmV2Cluster() *schema.Resource {
@@ -105,47 +106,6 @@ func resourceAwsCloudHsmV2Cluster() *schema.Resource {
 	}
 }
 
-func describeCloudHsmV2Cluster(conn *cloudhsmv2.CloudHSMV2, clusterId string) (*cloudhsmv2.Cluster, error) {
-	filters := []*string{&clusterId}
-	result := int64(1)
-	out, err := conn.DescribeClusters(&cloudhsmv2.DescribeClustersInput{
-		Filters: map[string][]*string{
-			"clusterIds": filters,
-		},
-		MaxResults: &result,
-	})
-	if err != nil {
-		log.Printf("[WARN] Error on retrieving CloudHSMv2 Cluster (%s) when waiting: %s", clusterId, err)
-		return nil, err
-	}
-
-	var cluster *cloudhsmv2.Cluster
-
-	for _, c := range out.Clusters {
-		if aws.StringValue(c.ClusterId) == clusterId {
-			cluster = c
-			break
-		}
-	}
-	return cluster, nil
-}
-
-func resourceAwsCloudHsmV2ClusterRefreshFunc(conn *cloudhsmv2.CloudHSMV2, clusterId string) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		cluster, err := describeCloudHsmV2Cluster(conn, clusterId)
-
-		if cluster == nil {
-			return 42, "destroyed", nil
-		}
-
-		if cluster.State != nil {
-			log.Printf("[DEBUG] CloudHSMv2 Cluster status (%s): %s", clusterId, *cluster.State)
-		}
-
-		return cluster, aws.StringValue(cluster.State), err
-	}
-}
-
 func resourceAwsCloudHsmV2ClusterCreate(d *schema.ResourceData, meta interface{}) error {
 	conn := meta.(*AWSClient).cloudhsmv2conn
 
@@ -158,60 +118,29 @@ func resourceAwsCloudHsmV2ClusterCreate(d *schema.ResourceData, meta interface{}
 		input.TagList = keyvaluetags.New(d.Get("tags").(map[string]interface{})).IgnoreAws().Cloudhsmv2Tags()
 	}
 
-	backupId := d.Get("source_backup_identifier").(string)
-	if len(backupId) != 0 {
-		input.SourceBackupId = aws.String(backupId)
+	if v, ok := d.GetOk("source_backup_identifier"); ok {
+		input.SourceBackupId = aws.String(v.(string))
 	}
 
 	log.Printf("[DEBUG] CloudHSMv2 Cluster create %s", input)
 
-	var output *cloudhsmv2.CreateClusterOutput
-
-	err := resource.Retry(180*time.Second, func() *resource.RetryError {
-		var err error
-		output, err = conn.CreateCluster(input)
-		if err != nil {
-			if isAWSErr(err, cloudhsmv2.ErrCodeCloudHsmInternalFailureException, "request was rejected because of an AWS CloudHSM internal failure") {
-				log.Printf("[DEBUG] CloudHSMv2 Cluster re-try creating %s", input)
-				return resource.RetryableError(err)
-			}
-			return resource.NonRetryableError(err)
-		}
-		return nil
-	})
-	if isResourceTimeoutError(err) {
-		output, err = conn.CreateCluster(input)
-	}
+	output, err := conn.CreateCluster(input)
 
 	if err != nil {
-		return fmt.Errorf("error creating CloudHSMv2 Cluster: %s", err)
+		return fmt.Errorf("error creating CloudHSMv2 Cluster: %w", err)
 	}
 
 	d.SetId(aws.StringValue(output.Cluster.ClusterId))
 	log.Printf("[INFO] CloudHSMv2 Cluster ID: %s", d.Id())
 	log.Println("[INFO] Waiting for CloudHSMv2 Cluster to be available")
 
-	targetState := cloudhsmv2.ClusterStateUninitialized
-	if len(backupId) > 0 {
-		targetState = cloudhsmv2.ClusterStateActive
-	}
-
-	stateConf := &resource.StateChangeConf{
-		Pending:    []string{cloudhsmv2.ClusterStateCreateInProgress, cloudhsmv2.ClusterStateInitializeInProgress},
-		Target:     []string{targetState},
-		Refresh:    resourceAwsCloudHsmV2ClusterRefreshFunc(conn, d.Id()),
-		Timeout:    d.Timeout(schema.TimeoutCreate),
-		MinTimeout: 30 * time.Second,
-		Delay:      30 * time.Second,
-	}
-
-	// Wait, catching any errors
-	_, errWait := stateConf.WaitForState()
-	if errWait != nil {
-		if len(backupId) == 0 {
-			return fmt.Errorf("[WARN] Error waiting for CloudHSMv2 Cluster state to be \"UNINITIALIZED\": %s", errWait)
-		} else {
-			return fmt.Errorf("[WARN] Error waiting for CloudHSMv2 Cluster state to be \"ACTIVE\": %s", errWait)
+	if input.SourceBackupId != nil {
+		if _, err := waiter.ClusterActive(conn, d.Id(), d.Timeout(schema.TimeoutCreate)); err != nil {
+			return fmt.Errorf("error waiting for CloudHSMv2 Cluster (%s) creation: %w", d.Id(), err)
+		}
+	} else {
+		if _, err := waiter.ClusterUninitialized(conn, d.Id(), d.Timeout(schema.TimeoutCreate)); err != nil {
+			return fmt.Errorf("error waiting for CloudHSMv2 Cluster (%s) creation: %w", d.Id(), err)
 		}
 	}
 
@@ -222,12 +151,30 @@ func resourceAwsCloudHsmV2ClusterRead(d *schema.ResourceData, meta interface{}) 
 	conn := meta.(*AWSClient).cloudhsmv2conn
 	ignoreTagsConfig := meta.(*AWSClient).IgnoreTagsConfig
 
-	cluster, err := describeCloudHsmV2Cluster(conn, d.Id())
+	cluster, err := finder.Cluster(conn, d.Id())
+
+	if err != nil {
+		return fmt.Errorf("error reading CloudHSMv2 Cluster (%s): %w", d.Id(), err)
+	}
 
 	if cluster == nil {
-		log.Printf("[WARN] CloudHSMv2 Cluster (%s) not found", d.Id())
+		if d.IsNewResource() {
+			return fmt.Errorf("error reading CloudHSMv2 Cluster (%s): not found after creation", d.Id())
+		}
+
+		log.Printf("[WARN] CloudHSMv2 Cluster (%s) not found, removing from state", d.Id())
 		d.SetId("")
-		return err
+		return nil
+	}
+
+	if aws.StringValue(cluster.State) == cloudhsmv2.ClusterStateDeleted {
+		if d.IsNewResource() {
+			return fmt.Errorf("error reading CloudHSMv2 Cluster (%s): %s after creation", d.Id(), aws.StringValue(cluster.State))
+		}
+
+		log.Printf("[WARN] CloudHSMv2 Cluster (%s) not found, removing from state", d.Id())
+		d.SetId("")
+		return nil
 	}
 
 	log.Printf("[INFO] Reading CloudHSMv2 Cluster Information: %s", d.Id())
@@ -276,29 +223,14 @@ func resourceAwsCloudHsmV2ClusterDelete(d *schema.ResourceData, meta interface{}
 		ClusterId: aws.String(d.Id()),
 	}
 
-	log.Printf("[DEBUG] CloudHSMv2 Delete cluster: %s", d.Id())
-	err := resource.Retry(180*time.Second, func() *resource.RetryError {
-		var err error
-		_, err = conn.DeleteCluster(input)
-		if err != nil {
-			if isAWSErr(err, cloudhsmv2.ErrCodeCloudHsmInternalFailureException, "request was rejected because of an AWS CloudHSM internal failure") {
-				log.Printf("[DEBUG] CloudHSMv2 Cluster re-try deleting %s", d.Id())
-				return resource.RetryableError(err)
-			}
-			return resource.NonRetryableError(err)
-		}
-		return nil
-	})
-	if isResourceTimeoutError(err) {
-		_, err = conn.DeleteCluster(input)
-	}
+	_, err := conn.DeleteCluster(input)
 
 	if err != nil {
-		return fmt.Errorf("error deleting CloudHSMv2 Cluster (%s): %s", d.Id(), err)
+		return fmt.Errorf("error deleting CloudHSMv2 Cluster (%s): %w", d.Id(), err)
 	}
 
-	if err := waitForCloudhsmv2ClusterDeletion(conn, d.Id(), d.Timeout(schema.TimeoutDelete)); err != nil {
-		return fmt.Errorf("error waiting for CloudHSMv2 Cluster (%s) deletion: %s", d.Id(), err)
+	if _, err := waiter.ClusterDeleted(conn, d.Id(), d.Timeout(schema.TimeoutDelete)); err != nil {
+		return fmt.Errorf("error waiting for CloudHSMv2 Cluster (%s) deletion: %w", d.Id(), err)
 	}
 
 	return nil
@@ -320,19 +252,4 @@ func readCloudHsmV2ClusterCertificates(cluster *cloudhsmv2.Cluster) []map[string
 		return []map[string]interface{}{certs}
 	}
 	return []map[string]interface{}{}
-}
-
-func waitForCloudhsmv2ClusterDeletion(conn *cloudhsmv2.CloudHSMV2, id string, timeout time.Duration) error {
-	stateConf := &resource.StateChangeConf{
-		Pending:    []string{cloudhsmv2.ClusterStateDeleteInProgress},
-		Target:     []string{cloudhsmv2.ClusterStateDeleted},
-		Refresh:    resourceAwsCloudHsmV2ClusterRefreshFunc(conn, id),
-		Timeout:    timeout,
-		MinTimeout: 30 * time.Second,
-		Delay:      30 * time.Second,
-	}
-
-	_, err := stateConf.WaitForState()
-
-	return err
 }
